refactor lang switcher
