+++ wp/wp-content/themes/phila.gov-theme/js/dev/main.js
@@ -22,3 +22,4 @@ require('./header-sticky')
 require('./parks-and-recreation')
 require('./gtm')
 require('./guides')
+require('./translations')
