+++ wp/wp-content/themes/phila.gov-theme/js/dev/phila-gov.js
@@ -182,16 +182,6 @@ module.exports = $(function () {
     });
   });
 
-  $(document).ready(function() {
-    $('#google_translate_element').bind('DOMNodeInserted', function() {
-        $('.goog-te-gadget .goog-te-gadget-simple span:first').replaceWith(function() { 
-            return "<a id='gte' role='menuitem'>More languages</>"; 
-        });
-    });
-    // "hard code" english translations label DD
-    $("#translate-english").text("English");
-    $("#translate-english-dropdown").text("English");
-
     //foundation equalizer rows
     //doesn't work with nested Equalizers, because a unique ID is required.
     if ($('.equal').length > 0) {
@@ -219,81 +209,4 @@ module.exports = $(function () {
     }
     var mainContent = $('.guide-content').eq(0);
 
-
-    //BEGIN Translation Bar
-
-    $("#lang-dropdown").on({
-      "show.zf.dropdown": function (e) {
-        $("html, body").css({
-          overflow: "hidden",
-          height: "100%",
-        });
-        toggleMenuOpen(true);
-      },
-      "hide.zf.dropdown": function (e) {
-        $("html, body").css({
-          overflow: "auto",
-          height: "auto",
-        });
-        toggleMenuOpen(false);
-      },
-    });
-
-    var hoverTimeout;
-
-    $('button#desktop-lang-button').hover(
-      function() {
-        clearTimeout(hoverTimeout);  
-        if(!$('#lang-dropdown').hasClass('is-open')){
-          $('#lang-dropdown').foundation('open');
-        }
-      },
-      function(e) {
-        if(e.relatedTarget.getAttribute('role') !== 'menuitem'){
-          hoverTimeout = setTimeout(function() {
-            $('#lang-dropdown').foundation('close');
-          }, 500);
-        } 
-      }
-    );
-
-    function toggleMenuOpen(isOpen) {
-      var $langDropdown = $("#lang-dropdown");
-      var $caretIcon = $(".translations-nav i.translate-caret");
-      var action = isOpen ? "addClass" : "removeClass";
-      $langDropdown.toggleClass("js-dropdown-active", isOpen);
-      $langDropdown.toggleClass("is-open", isOpen);
-      $caretIcon[action]("rotated");
-    }
-    
-    function setActiveLanguage(urlLanguage) {
-        $('#translations-menu> li').find('a').each(
-            function() {
-                if (urlLanguage === $.trim($(this).text())){
-                    $(this).addClass('active'); 
-                } else {
-                    $(this).removeClass('active');
-                }
-            });
-    }
-    function getUrlLanguage() {
-        var urlPath = windowPath.split('/');
-        var pathItem = urlPath[1];
-        var urlLanguage = "";
-        
-        switch(pathItem) {
-            case "zh":
-                urlLanguage = "中文";
-                break;
-            case "es":
-                urlLanguage = "Español";
-                break;
-            default:
-                urlLanguage = "English";
-        }
-        setActiveLanguage(urlLanguage);
-    }
-    getUrlLanguage();
-});
-//END Translation Bar
 });
